moved some misc rule classes into rules/options
