+++ lib/models/rules/options/combat_group_options.dart
