+++ lib/screens/unitSelector/unit_selection.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:gearforce/models/roster/roster.dart';
 import 'package:gearforce/models/rules/rule_set.dart';
-import 'package:gearforce/models/rules/special_unit_filter.dart';
+import 'package:gearforce/models/rules/options/special_unit_filter.dart';
 import 'package:gearforce/models/unit/role.dart';
 import 'package:gearforce/models/unit/unit.dart';
 import 'package:gearforce/screens/unitSelector/selected_unit_feedback.dart';
